pyproject.toml
[project]
name = "najork"
version = "0.1.0"
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Programming Language :: Python"
]
description = "A Kinetic Sequencer"
authors = [
	{ name="Mark Kennedy", email="mark@verynoisy.com" }
]

requires-python = ">=3.10"

dependencies = [
	"pycairo",
	"PyGObject",
	"Shapely < 1.9",
	"Jinja2 < 3.2",
	"PyYAML < 7.0",
	"py-expression-eval <0.4",
	"oscpy < 0.7.0",
	"python-rtmidi < 1.6",
	"mido < 1.3"
]

[project.optional-dependencies]
dev = [
	"pytest",
	"pytest-coverage",
	"pylint",
	"flake8",
	"ipython",
	"pytest-asyncio"
]
